gateway/requirements.txt
cassandra-driver
minio
pika
fastapi
uvicorn
python-multipart